Add widget to module migration script.
If a widget and a module with the same widget and it's patterns are in the same book, this script can update all the content to change references from the non-modular widget to the modular one, and delete the widget and associated non-modular patterns. Expected flow is:

(1) Someone creates a modular version of widget + patterns
(2) Module is uploaded to Module Repo
(3) Module is installed into projects from Repo / UI
(4) This script is run to cleanup all projects that now have two forms of the widget & patterns.

This script adds a repo dependency on the Beautiful Soup 4 HTML parsing library.